+++ Practica8/hello.c
@@ -1,5 +1,5 @@
 #include<stdio.h>
 int main(){
-    printf("Hola mundo...\n");
+    printf("Hola mundo.\n");
     return 0;
 }
